+++ app/src/src_nofdroid/org/breezyweather/sources/wmosevereweather/WmoSevereWeatherAlertWrapper.kt
@@ -1,0 +1,8 @@
+package org.breezyweather.sources.wmosevereweather
+
+import breezyweather.domain.weather.model.Alert
+
+data class WmoSevereWeatherAlertWrapper(
+    val alert: Alert,
+    val url: String?
+)