+++ maps-utils/src/main/java/com/google/maps/android/model/LatLng.kt
@@ -15,6 +15,18 @@ class LatLng(
         this.longitude = max(-180.0, min(180.0, longitude))
     }
 
+    override fun equals(other: Any?): Boolean {
+        return if (other is LatLng) {
+            latitude == other.latitude && longitude == other.longitude
+        } else false
+    }
+
+    override fun hashCode(): Int {
+        var result = latitude.hashCode()
+        result = 31 * result + longitude.hashCode()
+        return result
+    }
+
     override fun toString(): String {
         return "LatLng: [latitude=$latitude, longitude=$longitude]"
     }