+++ CHANGELOG.md
@@ -3,9 +3,16 @@
 **New features**
 - We now have our own repo for F-Droid and Google Play flavors of Breezy Weather. If you want to switch from our current repo (Standard flavor) to a different flavor, you should remove the repo for the Standard flavor first to avoid getting updates for the wrong flavor. [Link to Breezy Weather F-Droid repos](https://github.com/breezy-weather/fdroid-repo/blob/main/README.md)
 
+**Weather sources**
+- [World Meteorological Organization (WMO) Severe Weather] Added experimental support as secondary alert source. Translations are not supported (yet). Please report as soon as possible if you encounter a refresh error, making sure you mention the country of your location in the issue.
+
 **Improvements and fixes**
 - Background weather update will now only process the first location, unless you have multi cities usage. This makes the process more green and efficient, and other secondary locations will automatically be updated when you open them instead (this means that if you don’t check these locations very often, it will no longer uselessly refresh them in the background). A document has been written to describe in details how the full weather update process work: https://github.com/breezy-weather/breezy-weather/blob/main/docs/UPDATES.md
 - Fix AccuWeather being the default weather source instead of automatic recommendations
+- Fix some sources not being available for current location if current location didn’t change
+
+**Translations**
+- Translations updated
 
 
 # Version 5.1.3-beta (2024-03-13)